refactor: update routing in App component and standardize default constants for Dashboard
