+++ explore-classifiers/src/Components/DashboardPage.tsx
@@ -1,6 +1,6 @@
 import { useState } from "react";
 import Dashboard from "./Dashboard";
-import { DEFAULT_SUMMARY } from "../Constants/Common/defaultSummaryText";
+import { DEFAULT_SUMMARY } from "../Constants/Common/DashboardDefaults";
 
 export default function DashboardPage() {
     const [searchQuery, setSearchQuery] = useState(DEFAULT_SUMMARY);
