+++ explore-classifiers/src/Constants/Common/DashboardDefaults.ts
@@ -1,0 +1,3 @@
+export const DEFAULT_SUMMARY = 'This summary page shows results from all classifications.';
+
+export const DEFAULT_CANCER_TYPE = 'ZERO2';
