--- explore-classifiers/src/Constants/Common/defaultSummaryText.ts
@@ -1,1 +1,0 @@
-export const DEFAULT_SUMMARY = 'This summary page shows results from all classifications.';
